refactor(domain): split task entity into focused sibling modules
The 815-LOC task.ts was the largest domain entity. Split by concern into
sibling files under src/domain/entity/, with task.ts now hosting only the
type surface (143 LOC) — keeps the `Task` / `TodoTask` / `InProgressTask`
re-imports zero-diff for the 50+ pure-type consumers.

- task.ts (143) — types + re-exports from attempt
- task-factory.ts (187) — createTask + updateTask + criteria validation
- task-attempts.ts (204) — startNextAttempt + record/append/set helpers
- task-settle.ts (154) — markTaskDone / failCurrentAttempt / recordTaskEscalation
- task-lifecycle.ts (73) — markTaskBlocked / unblockTask / resetTaskToTodo / setTaskDependsOn
- task-graph.ts (109) — latestCritique / nextAvailableTask / validateTaskGraph

No barrels, no behaviour change, all files under 250 LOC.
