+++ src/business/task/commit-task.ts
@@ -2,7 +2,8 @@ import { Result } from '@src/domain/result.ts';
 import type { Logger } from '@src/business/observability/logger.ts';
 import type { SprintId } from '@src/domain/value/id/sprint-id.ts';
 import type { UpdateTask } from '@src/domain/repository/task/update-task.ts';
-import { type InProgressTask, recordRunningAttemptCommit } from '@src/domain/entity/task.ts';
+import type { InProgressTask } from '@src/domain/entity/task.ts';
+import { recordRunningAttemptCommit } from '@src/domain/entity/task-attempts.ts';
 import { type AbsolutePath } from '@src/domain/value/absolute-path.ts';
 import { CommitSha } from '@src/domain/value/commit-sha.ts';
 import type { InvalidStateError } from '@src/domain/value/error/invalid-state-error.ts';
