+++ src/business/task/escalation-policy.ts
@@ -2,7 +2,8 @@ import { Result } from '@src/domain/result.ts';
 import type { EventBus } from '@src/business/observability/event-bus.ts';
 import type { Logger } from '@src/business/observability/logger.ts';
 import { mergeEscalationMap } from '@src/business/task/escalation-map.ts';
-import { type InProgressTask, recordTaskEscalation } from '@src/domain/entity/task.ts';
+import type { InProgressTask } from '@src/domain/entity/task.ts';
+import { recordTaskEscalation } from '@src/domain/entity/task-settle.ts';
 import type { IsoTimestamp } from '@src/domain/value/iso-timestamp.ts';
 import type { InvalidStateError } from '@src/domain/value/error/invalid-state-error.ts';
 import type { ValidationError } from '@src/domain/value/error/validation-error.ts';
