+++ src/business/task/start-attempt.ts
@@ -4,7 +4,9 @@ import type { SprintId } from '@src/domain/value/id/sprint-id.ts';
 import type { FindTaskById } from '@src/domain/repository/task/find-task-by-id.ts';
 import type { UpdateTask } from '@src/domain/repository/task/update-task.ts';
 import type { RecoveryContext } from '@src/domain/entity/attempt.ts';
-import { failCurrentAttempt, startNextAttempt, type InProgressTask, type Task } from '@src/domain/entity/task.ts';
+import type { InProgressTask, Task } from '@src/domain/entity/task.ts';
+import { startNextAttempt } from '@src/domain/entity/task-attempts.ts';
+import { failCurrentAttempt } from '@src/domain/entity/task-settle.ts';
 import { InvalidStateError } from '@src/domain/value/error/invalid-state-error.ts';
 import type { NotFoundError } from '@src/domain/value/error/not-found-error.ts';
 import type { StorageError } from '@src/domain/value/error/storage-error.ts';
