+++ src/domain/entity/task-lifecycle.ts
@@ -1,0 +1,73 @@
+import { Result } from '@src/domain/result.ts';
+import type { BlockedTask, Task, TodoTask } from '@src/domain/entity/task.ts';
+import type { TaskId } from '@src/domain/value/id/task-id.ts';
+import { requireStatus } from '@src/domain/value/require-status.ts';
+import { InvalidStateError } from '@src/domain/value/error/invalid-state-error.ts';
+import { ValidationError } from '@src/domain/value/error/validation-error.ts';
+
+export const markTaskBlocked = (task: Task, reason: string): Result<BlockedTask, InvalidStateError> => {
+  const guard = requireStatus(
+    'task',
+    task,
+    ['todo', 'in_progress'] as const,
+    'mark-blocked',
+    'Done or already-blocked tasks cannot be re-blocked.'
+  );
+  if (!guard.ok) return Result.error(guard.error);
+  return Result.ok({ ...guard.value, status: 'blocked', blockedReason: reason });
+};
+
+export const unblockTask = (task: Task): Result<TodoTask, InvalidStateError> => {
+  const guard = requireStatus('task', task, ['blocked'] as const, 'unblock');
+  if (!guard.ok) return Result.error(guard.error);
+  const { blockedReason: _ignored, ...rest } = guard.value;
+  void _ignored;
+  return Result.ok({ ...rest, status: 'todo' });
+};
+
+/**
+ * Reset stale `in_progress` back to `todo` (for crash recovery). Requires there to be no
+ * unsettled running attempt — call `failCurrentAttempt(..., 'aborted')` first to settle it.
+ */
+export const resetTaskToTodo = (task: Task): Result<TodoTask, InvalidStateError> => {
+  if (task.status === 'todo') return Result.ok(task);
+  const guard = requireStatus(
+    'task',
+    task,
+    ['in_progress'] as const,
+    'reset-to-todo',
+    'Only `in_progress` tasks can be reset to todo.'
+  );
+  if (!guard.ok) return Result.error(guard.error);
+  const last = guard.value.attempts[guard.value.attempts.length - 1];
+  if (last !== undefined && last.status === 'running') {
+    return Result.error(
+      new InvalidStateError({
+        entity: 'task',
+        currentState: 'in_progress',
+        attemptedAction: 'reset-to-todo',
+        message: `task '${guard.value.id}' has a running attempt n=${last.n}`,
+        hint: 'Settle the attempt via failCurrentAttempt(..., "aborted") before resetting.',
+      })
+    );
+  }
+  return Result.ok({ ...guard.value, status: 'todo' });
+};
+
+/**
+ * Replace `dependsOn`. Rejects self-edges; deeper cycle detection (A→B→A) needs the
+ * full task graph and lives in `validateTaskGraph`.
+ * @public
+ */
+export const setTaskDependsOn = (task: Task, deps: readonly TaskId[]): Result<Task, ValidationError> => {
+  if (deps.includes(task.id)) {
+    return Result.error(
+      new ValidationError({
+        field: 'task.dependsOn',
+        value: deps,
+        message: `task '${task.id}' cannot depend on itself`,
+      })
+    );
+  }
+  return Result.ok({ ...task, dependsOn: [...deps] });
+};
