+++ tests/e2e/cli/task.test.ts
@@ -1,6 +1,6 @@
 import { afterEach, beforeEach, describe, expect, it } from 'vitest';
 import { createFsTaskRepository } from '@src/integration/persistence/task/repository.ts';
-import { markTaskBlocked } from '@src/domain/entity/task.ts';
+import { markTaskBlocked } from '@src/domain/entity/task-lifecycle.ts';
 import { makeDraftSprint, makeDoneTask, makeTodoTask } from '@tests/fixtures/domain.ts';
 import { createCliHome, runCliCaptured, type CliHome } from '@tests/e2e/cli/_harness.ts';
