+++ tests/integration/application/ui/tui/views/sprints-view.test.tsx
@@ -16,7 +16,8 @@ import { renderView } from '@tests/integration/application/ui/tui/_harness.tsx';
 import { createPromptQueue } from '@src/application/ui/tui/prompts/prompt-queue.ts';
 import { makeDraftSprint, makeTodoTask } from '@tests/fixtures/domain.ts';
 import { noopLogger } from '@tests/fixtures/noop-logger.ts';
-import { failCurrentAttempt, startNextAttempt } from '@src/domain/entity/task.ts';
+import { startNextAttempt } from '@src/domain/entity/task-attempts.ts';
+import { failCurrentAttempt } from '@src/domain/entity/task-settle.ts';
 import { IsoTimestamp } from '@src/domain/value/iso-timestamp.ts';
 
 const fakeSprintRepo = (sprints: readonly Sprint[]): SprintRepository =>
