+++ tests/unit/business/task/cancel-active-task.test.ts
@@ -1,7 +1,8 @@
 import { describe, expect, it } from 'vitest';
 import { Result } from '@src/domain/result.ts';
 import { cancelActiveTaskUseCase } from '@src/business/task/cancel-active-task.ts';
-import { markTaskBlocked, type BlockedTask, type Task } from '@src/domain/entity/task.ts';
+import type { BlockedTask, Task } from '@src/domain/entity/task.ts';
+import { markTaskBlocked } from '@src/domain/entity/task-lifecycle.ts';
 import type { UpdateTask } from '@src/domain/repository/task/update-task.ts';
 import type { SprintId } from '@src/domain/value/id/sprint-id.ts';
 import { StorageError } from '@src/domain/value/error/storage-error.ts';
